GUI / base class
Instance of class base is now an argument of the main GUI. It will be
used and adapted via entries of the GUI or the batch processing.
Very much in the style Guus had in mind before.

The gui is a singleton. It is possible to create the gui as a class and
not as a script, however combining this with case is not obvious.