made IO smarter
Several improvements to the IO class:
The load_thumb_frame function now looks if there is a result dataframe
it can use if a result thumb is asked for. if 'prior' is given as an
option to the function it will still load the thumbnail frame from a
prior location. The load_data_frame and load_sample function also look
in the results and load this when available. This makes it easy to
browse results and skips time intensive steps such as the creation of a
sample overview image.